requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
certifi==2019.9.11
cffi==1.12.3
cryptography==2.7
future==0.17.1
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pkg-resources==0.0.0
pycparser==2.19
pylint==2.3.1
python-telegram-bot==12.1.0
six==1.12.0
tornado==6.0.3
typed-ast==1.4.0
wrapt==1.11.2
